+++ orchestration/retirements/e04-codenexus-direct/rollback-rehearsal/run-code-intel.ps1
@@ -3511,7 +3511,7 @@ if (-not $SkipOpenSpec) {
             toolchainDigests = @(
                 "7fa18d2f751bc877c3367e314175e400c1a784a30fabc69b2a02efafcb6f3c85",
                 "25a2185026cb61771ff2e5f4c2364687d01158cfc9a8266d00a20e5573ba1bde",
-                "9fad4644b8cfb88fbec6f4603e7fcb099dc296fdf29286a6909279fab1450d0e"
+                "5090efd13c07531c249637d8e5857f0d13f3ecb8f0d02fb6e858747ea7d8c3d8"
             )
         }
         snapshot = $workflowSnapshot.snapshot